QASystemOnMedicalKGModified/requirements.txt
# # certifi @ file:///opt/conda/conda-bld/certifi_1655968806487/work/certifi
# charset-normalizer==2.1.0
# idna==3.3
# interchange==2021.0.4
# lxml==4.9.1
# monotonic==1.6
# packaging==21.3
# pansi==2020.7.3
# py2neo==2021.2.3
# pyahocorasick==1.4.4
# Pygments==2.12.0
# pymongo==4.1.1
# pyparsing==3.0.9
# pytz==2022.1
# requests==2.28.1
# six==1.16.0
# tqdm==4.64.0
# urllib3==1.26.10
charset-normalizer==2.1.0
idna==3.3
interchange==2021.0.4
lxml==4.9.1
monotonic==1.6
packaging==21.3
pansi==2020.7.3
py2neo==2021.2.4
pyahocorasick==1.4.4
Pygments==2.12.0
pymongo==4.1.1
pyparsing==3.0.9
pytz==2022.1
requests==2.28.1
six==1.16.0
tqdm==4.64.0
urllib3==1.26.10
